pyproject.toml
[project]
name = "citeiq"
version = "0.1.0"
description = "Reference normalization, quality scoring, and reporting tool."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "CiteIQ Team" }
]
dependencies = [
  "typer>=0.9.0",
  "pydantic>=2.5.0",
  "requests>=2.31.0",
  "pandas>=2.1.0",
  "rapidfuzz>=3.5.2",
  "scikit-learn>=1.3.0",
  "networkx>=3.1",
  "matplotlib>=3.8.0",
  "python-dateutil>=2.8.2",
  "orjson>=3.9.10",
  "bibtexparser>=1.4.1"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-mock",
  "ruff",
  "mypy"
]

[project.scripts]
citeiq = "citeiq.cli:app"

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"
